% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_valid_rescalers}
\alias{check_valid_rescalers}
\title{Check Valid Rescalers}
\usage{
check_valid_rescalers(x, x_name = substitute(x))
}
\arguments{
\item{x}{A character vector to test for valid rescalers.}

\item{x_name}{A string of the object name.}
}
\value{
An informative error if the check failed otherwise the original
  object.
}
\description{
Check Valid Rescalers
}
