% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rescaler_colnames}
\alias{get_rescaler_colnames}
\title{Get Rescaler Colnames}
\usage{
get_rescaler_colnames(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of the rescaler column names.
}
\description{
Get Rescaler Colnames
}
\examples{
get_rescaler_colnames(c("log(mean)*", "sqrt(cc)="))
}
