% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale-c.R
\name{rescale_c}
\alias{rescale_c}
\title{Rescale Transformed Data}
\usage{
rescale_c(data, data2 = data, colnames = character(0))
}
\arguments{
\item{data}{The data frame to rescale.}

\item{data2}{A data frame to use for the rescaling.}

\item{colnames}{A character vector of column names to transform and/or
rescale.}
}
\value{
The data frame with transformed and rescaled columns.
}
\description{
Rescales transformed columns in a data frame based on the transformed columns
in a second data frame. The columns are rescaled by subtracting values and
then dividing by values.
}
\details{
The column names can include a single function for the transform as well as
the following suffices: + (subtract mean), - (subtract min), =(subtract min
and add 1), / (divide by sd) and * (subtract mean and divide by sd).
}
\seealso{
\code{\link{rescale}}
}
