% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale-f.R
\name{rescale_f}
\alias{rescale_f}
\title{Rescale Transformed Data}
\usage{
rescale_f(
  data,
  data2 = data,
  transform = list(),
  subtract = list(),
  divide_by = list()
)
}
\arguments{
\item{data}{The data frame to rescale.}

\item{data2}{A data frame to use for the rescaling.}

\item{transform}{A named list where the name(s) indicate the function(s) to
use for the transformations and the elements indicate the columns to
transform.}

\item{subtract}{A named list where the name(s) indicate the function(s) to
use when determining the value to subtract and the elements indicate the
columns.}

\item{divide_by}{A named list where the name(s) indicate the function(s) to
use when determining the value to divide by and the elements indicate the
columns.}
}
\value{
The data frame with transformed and rescaled columns.
}
\description{
Rescales transformed columns in a data frame based on the transformed columns
in a second data frame. The columns are rescaled by subtracting values and
then dividing by values.
}
\examples{
rescale_f(datasets::mtcars,
  transform = list(log = "mpg"),
  subtract = list(mean = c("mpg", "disp"), min = "gear"),
  divide_by = list(sd = c("mpg", "hp")))
}
\seealso{
\code{\link{rescale}}
}
