% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale Data}
\usage{
rescale(data, data2 = data, center = character(0), scale = character(0))
}
\arguments{
\item{data}{The data frame to center and scale.}

\item{data2}{A data frame to use for the centering and scaling.}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}
}
\value{
The data frame with rescaled columns.
}
\description{
Centers and scales columns in a data frame based on the columns in a second
data frame. A column is centered by subtracting the mean and scaled by
dividing by the standard deviation. Columns in \code{scale} are automatically
added to \code{center} so they are standardised.
}
\examples{
rescale(datasets::mtcars, scale = "mpg")
}
\seealso{
\code{\link[base]{scale}}, \code{\link{rescale_f}} and \code{\link{rescale_c}}
}
