% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector}
\usage{
rescale(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Returns the original vector but rescale between  0 and 1
}
\description{
Rescale a vector
}
